fix casing on IlpError
